Squashed commit of the following:
commit c97f6697fd240028bfbf4997f9d208ba9807b484
Author: sjuri04 <sjuri04@hotmail.com>
Date:   Wed Jan 19 14:58:41 2022 -0300

    mas cambios

    mas cambios

    Modifique

    Le di funcion al boton atras en ingreso de usuario y se bloqueo los numeros en ingreso de usuario para los textos.

    Avances

    Avances en general codigo y fomrs

    _

    cosas

    Modificado

    Modificado

    Mas Modificaciones

    cambios en gral

    Cambios

    Cambios interfaz y botonera

    Interfaz

    INTERFAZ

    arreglos

    arreglos en gral

    Cambiosss

    cambiosss

commit 654f865cb820daa545916aff6bda13f66dce3a1d
Author: sjuri04 <sjuri04@hotmail.com>
Date:   Tue Jan 18 04:11:45 2022 -0300

    datos

    actualize sintaxis

+++ Veterinaria.Dominio/Socio.cs
@@ -4,7 +4,7 @@
     {
         public string Nombre { get; set; }
         public string Apellido { get; set; }
-        public int CuentaBancaria { get; set; }
+        public long CuentaBancaria { get; set; }
         public int Cedula { get; set; }
         public string Direccion { get; set; }
         public string Ciudad { get; set; }
